#23 Create separate view for rendering GameOptionsView ( @PiotrWR & @tomdworniczak )
